// src/shared/dataprovider/api/model/agentDTOV2.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AgentStatus } from './agentStatus';
import { AgentTypeEnum } from './agentTypeEnum';

export interface AgentDTOV2 { 
    AgentId?: number;
    AgentName?: string;
    AgentType?: AgentTypeEnum;
    Status?: AgentStatus;
    UpdatedAt?: Date;
}